Add files via upload
Added initial files that contain the first assets to be used to create for the rest of the game. Currently player 1 and 2's inputs are added (movement and pushing buttons) as well as looping music and custom classes that contain music and sound effects. The current music and sound effects are NOT final as they belong to their respective copyright owners.